Simplify hook
